_requirements.txt
pytorch==1.12.0
torchvision==0.13.0
jupyter==1.0.0
scikit-learn==1.0.2
tifffile==2021.7.2
pandas==1.4.2
hydra-core==1.2.0
tqdm==4.64.0
wandb==0.12.20